// src/tasks/types.ts
export interface Task<T> {
  name: string;
  read(blockHash: Hash, cached: Cached, api: ApiPromise): Promise<T>;
  write(blockNumber: BlockNumber, value: T): Promise<void>;
}

export interface Cached {
  events: EventRecord[];
  sessionIndex: SessionIndex;
}

export interface NomidotBlock {
  authoredBy: AccountId;
  hash: Hash;
  startDateTime: Moment;
}